requirements.txt
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
Cython==3.1.1
easydict==1.13
eval_type_backport==0.2.2
flatbuffers==25.2.10
fonttools==4.58.0
gitdb==4.0.12
GitPython==3.1.44
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
insightface==0.7.3
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
narwhals==1.41.0
networkx==3.2.1
numpy==2.0.2
onnx==1.18.0
onnxruntime==1.19.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
prettytable==3.16.0
protobuf==6.31.0
pyarrow==20.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
stringzilla==3.12.5
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2024.8.30
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
zipp==3.22.0
